backend/requirements.txt
alembic==1.13.1
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
validators==0.28.3
Werkzeug==3.0.3
flasgger
flask_cors
